Update README.md
The Grand Ensemble Multiverse is a Quantum Multiverse, in accordance with the Many Worlds Interpretation of Qunantum Mechanics, in which each World is contained within the Cosmic 11-dimensional spacetime Bulk of M-Theory.

The physical universe is a four-dimensional spacetime, which can interact, gravitionaly, with other nearby spacetime regions, known as Etheric Universes. In some cases, super-strings, known as Etheric Strings may be attached at each end to two different Branes.

The gravitional influence of Etheric Strings and Etheric Universes, account for the missing "Dark Matter" and "Dark Energy" within the physical universe.

An Etheric Body is formed through the gravitational influence of a physical body. It may have consciousness and continue to exist, even after the destruction of the physical body around which it was formed.